fix: :bug: Fix bugs in checkIfNeedArchived if algoliaIndexClient is null

+++ apollo/src/datasources/TagDataSource.js
@@ -479,7 +479,9 @@ class TagDataSource extends DataSource {
       await this.triggerEvent('archived', idWithResultData);
 
       // after archiving, need to delete the record in the index of algolia
-      await this.algoliaIndexClient.deleteObject(tagId);
+      if (this.algoliaIndexClient) {
+        await this.algoliaIndexClient.deleteObject(tagId);
+      }
     }
   }
